refactor tx ui component into own file
